Move union type constituents into the same file as the union type itself
