--- packages/core/src/types/external/AuditLogFilterBetween.ts
@@ -1,16 +1,0 @@
-/**
- * Match audit log entries captured between two timestamps.
- *
- * @public
- */
-export interface AuditLogFilterBetween {
-  /**
-   * Timestamp from which entries should be returned.
-   */
-  start: Date;
-
-  /**
-   * Timestamp before which entries should be returned.
-   */
-  end: Date;
-}
