--- packages/core/src/types/external/AuditLogFilterTail.ts
@@ -1,11 +1,0 @@
-/**
- * Match the last N audit log entries.
- *
- * @public
- */
-export interface AuditLogFilterTail {
-  /**
-   * The maximum number of entries to return.
-   */
-  tail: number;
-}
