--- packages/core/src/types/external/RolePolicy.ts
@@ -1,8 +1,0 @@
-import type { PolicyBase } from "./PolicyBase";
-import type { RolePolicyBody } from "./RolePolicyBody";
-
-/** @alpha */
-export interface RolePolicy extends PolicyBase {
-  /** @alpha */
-  rolePolicy: RolePolicyBody;
-}
