+++ packages/core/src/types/external/index.ts
@@ -2,9 +2,6 @@ export * from "./AccessLogEntry";
 export * from "./AddOrUpdatePoliciesRequest";
 export * from "./AddOrUpdateSchemasRequest";
 export * from "./AuditLogFilter";
-export * from "./AuditLogFilterBetween";
-export * from "./AuditLogFilterSince";
-export * from "./AuditLogFilterTail";
 export * from "./AuditTrail";
 export * from "./AuxData";
 export * from "./CheckInput";
@@ -20,23 +17,18 @@ export * from "./CheckResourcesResultResource";
 export * from "./Condition";
 export * from "./Constants";
 export * from "./DecisionLogEntry";
-export * from "./DecisionLogEntryCheckResources";
 export * from "./DecisionLogEntryMethod";
-export * from "./DecisionLogEntryPlanResources";
 export * from "./DecodedAuxData";
 export * from "./DeleteSchemasRequest";
 export * from "./DeleteSchemasResponse";
 export * from "./DerivedRoleDefinition";
-export * from "./DerivedRoles";
 export * from "./DerivedRolesBody";
 export * from "./DisablePoliciesRequest";
 export * from "./DisablePoliciesResponse";
 export * from "./Effect";
 export * from "./EnablePoliciesRequest";
 export * from "./EnablePoliciesResponse";
-export * from "./ExportConstants";
 export * from "./ExportConstantsBody";
-export * from "./ExportVariables";
 export * from "./ExportVariablesBody";
 export * from "./GetPoliciesRequest";
 export * from "./GetPoliciesResponse";
@@ -77,37 +69,24 @@ export * from "./PlanExpressionOperand";
 export * from "./PlanExpressionValue";
 export * from "./PlanExpressionVariable";
 export * from "./PlanKind";
-export * from "./PlanResourcesConditionalOutput";
-export * from "./PlanResourcesConditionalResponse";
 export * from "./PlanResourcesInput";
 export * from "./PlanResourcesMetadata";
 export * from "./PlanResourcesOutput";
-export * from "./PlanResourcesOutputBase";
 export * from "./PlanResourcesRequest";
-export * from "./PlanResourcesRequestBase";
-export * from "./PlanResourcesRequestWithAction";
-export * from "./PlanResourcesRequestWithActions";
 export * from "./PlanResourcesResponse";
-export * from "./PlanResourcesResponseBase";
-export * from "./PlanResourcesUnconditionalOutput";
-export * from "./PlanResourcesUnconditionalResponse";
 export * from "./Policy";
-export * from "./PolicyBase";
 export * from "./PolicyMetadata";
 export * from "./Principal";
-export * from "./PrincipalPolicy";
 export * from "./PrincipalPolicyBody";
 export * from "./PrincipalRule";
 export * from "./PrincipalRuleAction";
 export * from "./ReloadStoreRequest";
 export * from "./Resource";
 export * from "./ResourceCheck";
-export * from "./ResourcePolicy";
 export * from "./ResourcePolicyBody";
 export * from "./ResourceQuery";
 export * from "./ResourceRule";
 export * from "./ResourceSearch";
-export * from "./RolePolicy";
 export * from "./RolePolicyBody";
 export * from "./RoleRule";
 export * from "./Schema";
